+++ src/ios/app.m
@@ -36,7 +36,10 @@ void ctd_emit_control(ctd_handle target, id sender) {
     NSString *text = nil;
 
     int32_t made_as = ctd_slot_kind(target);
-    if ([sender isKindOfClass:[UISlider class]]) {
+    if ([sender isKindOfClass:[UIStepper class]]) {
+        kind = CTD_EV_VALUE_CHANGED;
+        index = (int64_t)[(UIStepper *)sender value];
+    } else if ([sender isKindOfClass:[UISlider class]]) {
         kind = CTD_EV_VALUE_CHANGED;
         index = (int64_t)[(UISlider *)sender value];
     } else if ([sender isKindOfClass:[UISwitch class]]) {